backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pdfplumber==0.11.7
pandas==2.2.3
openpyxl==3.1.2
openai==1.46.0
sentence-transformers==2.7.0
faiss-cpu==1.12.0
networkx==3.3
sqlalchemy==2.0.34
python-multipart==0.0.9
python-dotenv==1.0.1
numpy==1.24.3
spacy==3.7.2
google-api-python-client==2.108.0
google-auth-oauthlib==1.1.0
neo4j==5.14.1
pypdf2==3.0.1
langchain==0.1.0
tiktoken==0.5.2
beautifulsoup4==4.12.2
transformers==4.36.2
torch==2.1.2

